src/WebApiContrib.Formatting.CollectionJson/CollectionJsonFormatter.cs
using System;
using System.Net.Http.Formatting;
using System.Net.Http.Headers;
using Newtonsoft.Json;
using Newtonsoft.Json.Serialization;

namespace WebApiContrib.Formatting.CollectionJson
{
    public class CollectionJsonFormatter : JsonMediaTypeFormatter
    {
        public CollectionJsonFormatter()
        {
            SupportedMediaTypes.Clear();
            SupportedMediaTypes.Add(new MediaTypeHeaderValue("application/vnd.collection+json"));
            SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
            SerializerSettings.Formatting = Newtonsoft.Json.Formatting.Indented;
            SerializerSettings.ContractResolver =
                new CamelCasePropertyNamesContractResolver();
        }

        public override bool CanWriteType(Type type)
        {
            return (type == typeof (ReadDocument) || type == typeof (WriteDocument));
        }

        public override bool CanReadType(Type type)
        {
            return (type == typeof (ReadDocument) || type == typeof (WriteDocument));
        }
    }
}